Move lib/selinux to session/selinux
